add prompts for user choice
Instead of relying on the user to supply their repo type on the command
line and prompting them for a [Y/N] choice, I have used the pick API to
present the user with a prompt and navigable list of choices.

This eliminates the need to supply command line args for the time being,
and removes a lot of boilerplate code. I have also suppressed all git
output.